Changed the application window so that it is not resizable.
Updated the Java wrapper and added a Utility class with some functions to cast between classes in the GameObject hierarchy. This is because Java does not support multiple inheritance, and therefore some of the wrapped Java classes do not have all of the functions they should have, and you cannot always cast to the correct class since in Java it is not part of the hierarchy (due to the non-MI).

In future we should consider creating new SWIG interface files that mirror the source header files instead of blindly processing the C++ header files directly. Then we can explicitly control what functions are available and handle these MI problems more elegantly.